requirements.txt
dify_plugin>=0.5.0
python-docx>=1.1.2
